+++ packages/components/src/ui/navigation-menu.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import * as NavigationMenuPrimitive from "@radix-ui/react-navigation-menu";
-import { ChevronDown } from "lucide-react";
+import { Icon } from "../icon/index.js";
 import { cva } from "class-variance-authority";
 import { cn } from "../utils.js";
 
@@ -47,7 +47,8 @@ const NavigationMenuTrigger = React.forwardRef<
     {...props}
   >
     {children}{" "}
-    <ChevronDown
+    <Icon
+      name="chevronDown"
       className="relative top-px ml-1 size-3 transition duration-300 group-data-[state=open]:rotate-180"
       aria-hidden="true"
     />