+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @arcevo/facet-docs
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - @arcevo/facet-components@1.0.2
+  - @arcevo/facet-auth@1.0.2
+  - @arcevo/facet-layout@1.0.2
+
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies [d94a724]
+  - @arcevo/facet-auth@1.0.1
+  - @arcevo/facet-components@1.0.1
+  - @arcevo/facet-layout@1.0.1
+  - @arcevo/facet-sdk@1.0.1
+  - @arcevo/facet-tokens@1.0.1
+
 ## 1.0.0
 
 ### Major Changes