+++ packages/components/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @arcevo/facet-components
 
+## 1.0.2
+
+### Patch Changes
+
+- feat: wire internal components through the Icon registry; Icon spreads SVG props for pass-through overrides
+- fix: navbar mobile menu closes when an item is tapped (custom mobileMenu included)
+
+## 1.0.1
+
+### Patch Changes
+
+- d94a724: chore: update homepage to facet.arcevocirqle.com.ng
+- Updated dependencies [d94a724]
+  - @arcevo/facet-tokens@1.0.1
+
 ## 1.0.0
 
 ### Major Changes