+++ packages/tokens/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @arcevo/facet-tokens
 
+## 1.0.1
+
+### Patch Changes
+
+- d94a724: chore: update homepage to facet.arcevocirqle.com.ng
+
 ## 1.0.0
 
 ### Major Changes